24 game algorithm v0.2
Submitted 1/29/2018 - This is a very WIP version of the number 4 algorithm. It works currently to check four numbers with every combination of operators (+, -, /, *). This gives all 64 possible numbers that can be made from the original four, but only when the are in a set order, in this case 4, 6, 8, 10. I do not have much time this week, but next week I will try to add a section to check the end result against 24, and move the order of the starting numbers.  I also have a v0.1 file if people want to see it, but it does not run correctly.